+++ src/afterfact.rs
@@ -424,7 +424,8 @@ fn _get_timestamp(time: &DateTime<Utc>) -> i64 {
     if configs::CONFIG.read().unwrap().args.is_present("utc") {
         time.timestamp()
     } else {
-        time.with_timezone(&Local).timestamp()
+        let offset_sec = Local.timestamp(0, 0).offset().local_minus_utc();
+        offset_sec as i64 + time.with_timezone(&Local).timestamp()
     }
 }