+++ src/detections/configs.rs
@@ -84,8 +84,8 @@ fn build_app<'a>() -> ArgMatches<'a> {
     -u --update-rules 'Update to the latest rules in the hayabusa-rules github repository.'
     -m --min-level=[LEVEL] 'Minimum level for rules. (Default: informational)'
     -l --live-analysis 'Analyze the local C:\\Windows\\System32\\winevt\\Logs folder (Windows Only. Administrator privileges required.)'
-    --start-timeline=[STARTTIMELINE] 'Start time of the event logs to load. (Example: '2018/11/28 12:00:00 +09:00')'
-    --end-timeline=[ENDTIMELINE] 'End time of the event logs to load. (Example: '2018/11/28 12:00:00 +09:00')'
+    --start-timeline=[STARTTIMELINE] 'Start time of the event logs to load. (Example: \"2018-11-28 12:00:00 +09:00\")'
+    --end-timeline=[ENDTIMELINE] 'End time of the event logs to load. (Example: \"2021-11-28 12:00:00 +09:00\")'
     --rfc-2822 'Output date and time in RFC 2822 format. (Example: Mon, 07 Aug 2006 12:34:56 -0600)'
     --rfc-3339 'Output date and time in RFC 3339 format. (Example: 2006-08-07T12:34:56.485214 -06:00)'
     -U --utc 'Output time in UTC format. (Default: local time)'
@@ -173,6 +173,7 @@ fn load_target_ids(path: &str) -> TargetEventIds {
 
 #[derive(Debug, Clone)]
 pub struct TargetEventTime {
+    parse_success_flag: bool,
     start_time: Option<DateTime<Utc>>,
     end_time: Option<DateTime<Utc>>,
 }
@@ -185,18 +186,20 @@ impl Default for TargetEventTime {
 
 impl TargetEventTime {
     pub fn new() -> Self {
+        let mut parse_success_flag = true;
         let start_time =
             if let Some(s_time) = CONFIG.read().unwrap().args.value_of("start-timeline") {
                 match DateTime::parse_from_str(s_time, "%Y-%m-%d %H:%M:%S %z") // 2014-11-28 21:00:09 +09:00
                 .or_else(|_| DateTime::parse_from_str(s_time, "%Y/%m/%d %H:%M:%S %z")) // 2014/11/28 21:00:09 +09:00
             {
                 Ok(dt) => Some(dt.with_timezone(&Utc)),
-                Err(err) => {
+                Err(_) => {
                     AlertMessage::alert(
                         &mut BufWriter::new(std::io::stderr().lock()),
-                        &format!("start-timeline field: {}", err),
+                        "start-timeline field: the timestamp format is not correct.",
                     )
                     .ok();
+                    parse_success_flag = false;
                     None
                 }
             }
@@ -208,29 +211,36 @@ impl TargetEventTime {
             .or_else(|_| DateTime::parse_from_str(e_time, "%Y/%m/%d %H:%M:%S %z")) // 2014/11/28 21:00:09 +09:00
         {
             Ok(dt) => Some(dt.with_timezone(&Utc)),
-            Err(err) => {
+            Err(_) => {
                     AlertMessage::alert(
                         &mut BufWriter::new(std::io::stderr().lock()),
-                        &format!("end-timeline field: {}", err),
+                        "end-timeline field: the timestamp format is not correct.",
                     )
                     .ok();
+                    parse_success_flag = false;
                     None
                 }
             }
         } else {
             None
         };
-        Self::set(start_time, end_time)
+        Self::set(parse_success_flag, start_time, end_time)
     }
 
     pub fn set(
+        input_parse_success_flag: bool,
         input_start_time: Option<chrono::DateTime<chrono::Utc>>,
         input_end_time: Option<chrono::DateTime<chrono::Utc>>,
     ) -> Self {
         Self {
+            parse_success_flag: input_parse_success_flag,
             start_time: input_start_time,
             end_time: input_end_time,
         }
+    }
+
+    pub fn is_parse_success(&self) -> bool {
+        self.parse_success_flag
     }
 
     pub fn is_target(&self, eventtime: &Option<DateTime<Utc>>) -> bool {
@@ -450,7 +460,7 @@ mod tests {
     fn target_event_time_filter() {
         let start_time = Some("2018-02-20T12:00:09Z".parse::<DateTime<Utc>>().unwrap());
         let end_time = Some("2020-03-30T12:00:09Z".parse::<DateTime<Utc>>().unwrap());
-        let time_filter = configs::TargetEventTime::set(start_time, end_time);
+        let time_filter = configs::TargetEventTime::set(true, start_time, end_time);
 
         let out_of_range1 = Some("1999-01-01T12:00:09Z".parse::<DateTime<Utc>>().unwrap());
         let within_range = Some("2019-02-27T01:05:01Z".parse::<DateTime<Utc>>().unwrap());
@@ -465,7 +475,7 @@ mod tests {
     fn target_event_time_filter_containes_on_time() {
         let start_time = Some("2018-02-20T12:00:09Z".parse::<DateTime<Utc>>().unwrap());
         let end_time = Some("2020-03-30T12:00:09Z".parse::<DateTime<Utc>>().unwrap());
-        let time_filter = configs::TargetEventTime::set(start_time, end_time);
+        let time_filter = configs::TargetEventTime::set(true, start_time, end_time);
 
         assert!(time_filter.is_target(&start_time));
         assert!(time_filter.is_target(&end_time));