+++ CHANGELOG.md
@@ -17,7 +17,8 @@
 
 **Bug Fixes:**
 
-- xxx
+- Fixed the `--start-timeline` and `--end-timeline` options as they were not working. (#546) @hitenkoku)
+
 
 ## v1.2.2 [2022/05/20]